fix: check it map loaded and load it call init handler without adding subscription.
The reason was in onInit hook setup process - there was a small time gap between map instance created and we pass link of instance to other places in our application where eventListener for 'load' event adds. If load event happens in this gap - subscribers that will be added after that just wait load event that not happens never again.