Ignore e2e and config changes in CI workflows
This commit updates the GitHub Actions workflow files to ignore changes
in 'e2e/' directory and 'playwright.config.ts' file. This is done to
prevent unnecessary CI runs when changes are made to end-to-end tests
or the Playwright configuration. Additionally, the commit standardizes
the use of single quotes for consistency across the workflow files.